requirements.txt
pikepdf>=8.0.0
pdfplumber>=0.10.0
llama-cpp-python>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pywebview>=5.0
apscheduler>=3.10.0
